refactored auth action'
