+++ src/reducers/index.js
@@ -1,7 +1,9 @@
-import { combineReducers } from 'redux'
-import userReducer         from './userReducer'
+import { combineReducers }   from 'redux'
+import userReducer           from './userReducer'
+import authenticationReducer from './authenticationReducer'
 
 const allReducers = combineReducers({
+  authentication: authenticationReducer,
   user: userReducer
 })
